refactor: move mm_* from request to multimodal.
