+++ xllm/core/distributed_runtime/vlm_master.cpp
@@ -28,8 +28,8 @@ limitations under the License.
 #include "core/common/message.h"
 #include "framework/chat_template/jinja_chat_template.h"
 #include "framework/model/model_args.h"
-#include "framework/request/mm_data.h"
-#include "framework/request/mm_input.h"
+#include "framework/multimodal/mm_data.h"
+#include "framework/multimodal/mm_input.h"
 #include "framework/request/request.h"
 #include "models/model_registry.h"
 #include "runtime/xservice_client.h"
