+++ xllm/core/distributed_runtime/vlm_master.h
@@ -29,7 +29,7 @@ limitations under the License.
 #include "common/types.h"
 #include "engine.h"
 #include "framework/chat_template/jinja_chat_template.h"
-#include "framework/request/mm_input.h"
+#include "framework/multimodal/mm_input.h"
 #include "framework/request/request_output.h"
 #include "framework/request/request_params.h"
 #include "master.h"
