+++ xllm/core/framework/CMakeLists.txt
@@ -17,6 +17,7 @@ add_subdirectory(model)
 add_subdirectory(parallel_state)
 add_subdirectory(prefix_cache)
 add_subdirectory(request)
+add_subdirectory(multimodal)
 add_subdirectory(sampling)
 add_subdirectory(state_dict)
 add_subdirectory(tokenizer)
