+++ xllm/core/framework/batch/batch.h
@@ -24,7 +24,7 @@ limitations under the License.
 #include <vector>
 
 #include "framework/batch/batch_forward_type.h"
-#include "framework/request/mm_data.h"
+#include "framework/multimodal/mm_data.h"
 #include "framework/request/request.h"
 #include "framework/request/sequence.h"
 #include "framework/request/sequences_group.h"
