+++ xllm/core/framework/batch/rec_batch_input_builder.h
@@ -22,7 +22,7 @@ limitations under the License.
 #include <vector>
 
 #include "framework/model/model_args.h"
-#include "framework/request/mm_data.h"
+#include "framework/multimodal/mm_data.h"
 #include "framework/request/rec_type.h"
 #include "framework/request/sequences_group.h"
 #include "runtime/forward_params.h"
