+++ xllm/core/framework/multimodal/mm_batch_data.h
