+++ xllm/core/framework/multimodal/mm_embedding_handler.h
