+++ xllm/core/framework/multimodal/mm_item_state.cpp
