+++ xllm/core/framework/multimodal/mm_type.cpp
