+++ xllm/core/runtime/forward_params.h
@@ -30,8 +30,8 @@ limitations under the License.
 #include "common/types.h"
 #include "framework/config/execution_config.h"
 #include "framework/model/model_input_params.h"
-#include "framework/request/mm_batch_data.h"
-#include "framework/request/mm_data.h"
+#include "framework/multimodal/mm_batch_data.h"
+#include "framework/multimodal/mm_data.h"
 #include "framework/sampling/beam_searcher.h"
 #include "framework/sampling/sampling_params.h"
 #include "platform/device.h"
