+++ xllm/core/runtime/mtp_worker_impl.cpp
@@ -23,7 +23,7 @@ limitations under the License.
 #include "core/framework/config/disagg_pd_config.h"
 #include "core/framework/config/kernel_config.h"
 #include "core/framework/config/speculative_config.h"
-#include "framework/request/mm_data.h"
+#include "framework/multimodal/mm_data.h"
 #include "spec_input_builder.h"
 #include "util/env_var.h"
 #include "util/pretty_print.h"
