+++ xllm/models/dit/pipelines/pipeline_longcat_image_edit.h
@@ -25,11 +25,11 @@
 
 #include "core/framework/dit_model_loader.h"
 #include "core/framework/model_context.h"
+#include "core/framework/multimodal/mm_batch_data.h"
+#include "core/framework/multimodal/mm_input.h"
 #include "core/framework/parallel_state/parallel_args.h"
 #include "core/framework/parallel_state/process_group.h"
 #include "core/framework/request/dit_request_state.h"
-#include "core/framework/request/mm_batch_data.h"
-#include "core/framework/request/mm_input.h"
 #include "core/framework/tokenizer/tokenizer.h"
 #include "core/layers/common/attention_metadata_builder.h"
 #include "core/layers/cuda/flashinfer_workspace.h"
