+++ xllm/processors/image_processor.h
@@ -20,8 +20,8 @@ limitations under the License.
 #include <vector>
 
 #include "core/framework/model/model_args.h"
-#include "core/framework/request/mm_data.h"
-#include "core/framework/request/mm_input.h"
+#include "core/framework/multimodal/mm_data.h"
+#include "core/framework/multimodal/mm_input.h"
 
 namespace xllm {
